Preliminary documentation
If the current API were final, this documentation would probably be
enough.  Unfortunately, the API still doesn't work for some important
use cases, so it'll have to change.  Thus the docs will change too.
This was mostly to get a flavor for Haddock and to remind myself that
documentation is necessary before releasing to Hackage.