Bitcoin API arguments are JSON values
The Bitcoin API fails if you pass it an argument of the wrong type.  For
instance, "getbalance" accepts two optional arguments: a string and an
integer.  If you pass the integer argument encoded as a string, the API
returns an error message about a type mismatch.  In order to encode the
necessary types, our API uses a list of JSON values.  This results in
code like:

    callBTC "getbalance" [ "", Number 0 ]

to find the unconfirmed balance of the "" account. At some point, it
might be convenient to provide a more specific function for each
possible API call.  Maybe something like,

    btcGetBalance "" 0

but that's not implemented yet and it would have to support optional
arguments so end up looking like,

    btcGetBalance (Just "") (Just 0)

which is fairly ugly.  It's also not entirely true since the first
argument is mandatory if a second argument is provided.  Obviously, I
need more time to figure out how this will work.