Better data types for validateAddress
The data type AddressValidation represents the information returned
from validateaddress RPC call.  Because it's a record, it can safely
grow to accommodate new information as the RPC changes.

Incidentally, this approach gives a cleaner implementation for
validateAddress which is always pleasant.

Constructors for AddressValidation are not exported as users should
never need to build one of these data structures themselves.