Move address type to a separate module
This allows me to easily release the address module as an independent
module if that ever becomes desirable.  I can imagine software that
generates new addresses or interacts directly with the Bitcoin network
wanting to handle addresses but not wanting to use the JSON API.

I also removed "Bitcoin" from the type names.  If there are naming
conflicts, users can import qualified.