Better conversion of Double to Amount
Converting a Double to an Amount using `fromRational . toRational`
retains floating point errors because of how toRational chooses the
denominator from a Double.  Since we know that Double's representing a
Bitcoin amount have the resolution of 1 Satoshi, we can convert to a
Ratio more intelligently.

This makes the test suite pass